requirements.txt
absl-py==1.4.0
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
audioread==3.0.0
axios==0.4.0
beautifulsoup4==4.12.2
billiard==4.1.0
blis==0.7.9
boilerplate==1.2.4
cachetools==5.3.1
catalogue==2.0.8
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.7
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
filelock==3.12.2
flatbuffers==23.5.26
fsspec==2023.6.0
gast==0.4.0
gdown==4.7.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.15.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
jax==0.4.12
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
kenlm==0.1
keras==2.12.0
kombu==5.3.1
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.40.1rc1
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pathy==0.10.1
pooch==1.6.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.23.3
punctuator==0.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycorrector==0.5.0
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyJWT==1.7.1
pypinyin==0.49.0
PySocks==1.7.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
Theano==1.0.5
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
typer==0.7.0
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
